Project Setup And CucumberExpressions Parser
This project is to contain all the helper libraries in order to realize
`ex_cucumber`:
* `Utils` houses all that shares between several apps
* `Playground` is to keep some experiments of interest in version control
* `CucumberExpressions` as a helper library
* `ExCucumber` as main

`ExCucumber` is to run `Gherkin Feature`-files. This requires
associating functionality to be run in the order as the various lines
appear in such feature files.
For instance, given the following feature file:

```gherkin
Feature: Calculator
    Scenario: Addition
        Given the first number: 5
        And a second number: 3
        Then their summation is: 8
```

On the developer side, we need to have a way of executing those steps by
associating some functionality to them. It would be a rather diminutive
developer experience if that association is to be made by typing out
those lines verbatim as they appear:

```elixir
defgiven "the first number: 5" do
    {:ok, %{first_number: 5}}
end
```

For this reason, traditional `Cucumber` would allow to abstract away by
means of `Regex` so we could express the association under `defgiven`
like this:
```elixir
defgiven ~r/the first number: (\d+)/ do |regex_match|
    [first_number] = regex_match
    {:ok, %{first_number: 5}}
end
```

However, `Regex` may allow us to abstract away but this comes at the
cost of readability. For this purpose, `CucumberExpressions` have been
introduced where instead we could for instance write as follows:
```elixir
defgiven "the first number: {int}" do |matches|
    {:ok, %{first_number: matches.int}}
end
```

The above syntax is for illustration purposes only, what counts is the
underlying idea of enabling the ability to associate functionality, e.g.:
```
do |matches|
    {:ok, %{first_number: matches.int}}
end
```
with a specific line in a `Gherkin Feature`-file, e.g.:
```
        Given the first number: 5
```
in a readable manner. This is the purpose of `CucumberExpressions`.
Kindly consult `@moduledoc` for further details.