Introduce Cucumber Expressions Matcher
Once we are able to parse cucumber expressions, we need a way to match
them. Using the parse tree made available to us with the work done
previously, we can linearly scan the sentence as it occurs in a feature
file and word by word arrive at the cucumber expression so that we can
ultimately access the functionality associated therewith.

In the meanwhile, we need to leverage the various features that
Cucumber Expressions provide us with:
  * Optionals
  * Alternatives
  * Parameter Types

The work previously done has introduced the concepts thereof, whereas
this commit was required to venture deeper into this territory.
Parameter Types is thus a main concern on part of the Matcher that needs
to relate them accordingly so that we can extract certain types of
information from a sentence as it occurs in a feature file; transform
it and exercise a set of validations on it.