Introduce Gherkin Macros
`ExCucumber` represents the front end and has its tentacles spread out
into:
  * `ExGherkin`
  * `CucumberExpressions`

`ExGherkin` comes with its own set of baggage: `gherkin-languages.*`. It
is used to parse feature files into `AST` so we can traverse accordingly

In order to be able to leverage `CucumberExpressions` during such
traversal; there is a need for a syntactic tool that a developer can
take recourse to that:
  * associates functionality with such expressions
  * references the feature file to be traversed
with a minimum of ceremony.

In the `Elixir`-community, we predominantly feel that syntax over and
above solving a problem needs to be clean and contribute to developer
happiness; while still conveying an overal `Elixir`-feel to its flow.
For this reason, 2 variants have been introduced to facilitate the
community to decide which one makes more sense. These variants are:
  * Module-based: e.g.: `Given._`, `When._`, etc..
  * Definition-based: e.g.: `defgiven`, `defwhen`, etc..

The module-based variant has the advantage of starting with a capital
letter which makes it closer to the way feature files are written, at
the expense of a seemingly nonsensical `_` that we delegate
functionality to. Despite its aesthetically pleasing outlook, especially
considering that there is already syntactic highlighting for `modules`;
it is not the typical `Elixir` way to leverage `macros`.

The definition-based variant represents the more traditional `Elixir`-way
instead.

These verbage styles are mutually exclusive of each other and which one
may choose can be configured together with a set of other options that
allow this tool to contribute to `Developer Happiness`. One of the
fundamental drives of this tool is the understanding that the BDD-philosophy
as espoused by `Cucumber` has its learning curve and that accordingly
documentation should be readily available at the time of need.

For this reason, another experimental deviation from the typical `Elixir`
way of doing things has been introduced; namely that one can configure
`errors` to have the verbosity of documentation. This can go rather
detailed and should especially assist junior developers at the time of
trouble shooting.

Besides the primary purpose above, there is a component of forcing this
project to get to the heart of the matter in terms of developer happiness.
The ritual of writing such elaborate, dynamic documentation should
assist in driving this. Ultimately, the objective is to move all this
detail somewhere online instead and to have error messages communicate
a link to access the same.